NSBenefits/ClientApp/src/app/cost/chart/chart.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { ChartComponent } from './chart.component';

describe('ChartComponent', () => {
  let component: ChartComponent;
  let fixture: ComponentFixture<ChartComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ ChartComponent ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(ChartComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should calculate the benefits cost percentage', () => {
    const result = component.getCostPercentage(10, 100);
    expect(result).toBe(10);
  });

  it('should calculate the bar width', () => {
    const result = component.getCostBarWidth(3);
    expect(result).toBe('30px');
  });
});
